Fix various shellcheck errors

+++ bin/build-trustedkeys.sh
@@ -1,20 +1,19 @@
 #!/usr/bin/env sh
 set -e
 
-if [ -z "$1" -o ! -e "$1" ]; then
+if [ -z "$1" ] || [ ! -e "$1" ]; then
     >&2 echo "Usage: build-trustedkeys.sh <public key file>"
     exit 1
 fi
 
-if [ -e trustedkeys.gpg ]; then
-  rm trustedkeys.gpg
-fi
+[ -e trustedkeys.gpg ] && rm trustedkeys.gpg
 # File needs to exists, otherwise Dockers bind mount will create a directory
 touch trustedkeys.gpg
 
+# shellcheck disable=SC1004
 docker run --rm \
-  -v "`pwd`/trustedkeys.gpg:/trustedkeys.gpg" \
-  -v "`pwd`/$1:/pubkey.asc:ro" \
+  -v "$(pwd)/trustedkeys.gpg:/trustedkeys.gpg" \
+  -v "$(pwd)/$1:/pubkey.asc:ro" \
   alpine sh -c '
   set -eu; \
   apk --no-cache add gnupg && \
